Conflicts:
	Assets/Prefabs/GUICamera.prefab
	Assets/Prefabs/prefab_pile_of_leifs.prefab
	Assets/Prefabs/prefab_player.prefab
	Assets/Scenes/test_johannes.unity
	Assets/Scripts/GUI/GUIControl.cs

+++ Assets/Scripts/Powerup/PowerupManager.cs
@@ -94,7 +94,7 @@ public class PowerupManager : MonoBehaviour {
 	 * Add the TimeBomb buff to the colliding player
 	 */
 	[RPC]
-	public void TimeBombGet(int playerID, int duration)
+	public void TimeBombGet(int playerID, float duration)
 	{
 		SyncMovement syncMovement;
 		GameObject player;
